requirements.txt
streamlit==1.32.0
openai==1.47.0
python-dotenv==1.0.0
instructor==1.0.0
ffmpeg-python==0.2.0
pydub==0.25.1
qdrant-client==1.7.0
python-docx==1.1.0
PyPDF2==3.0.1
